lab3/budget/pyproject.toml
[project]
name = "lab2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dishka>=1.5.0",
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
